# docs/interfaces/openapi-data-analyzation.yaml
openapi: 3.0.3
info:
  title: Darkwebsearch - Data Analyzer API
  version: "1.0.0"
  description: API for analyzing data and creating tags in Darkwebsearch.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: "/"
    description: Relative server URL

paths:
  /analyze:
    post:
      summary: Analyze data and create tags
      operationId: analyzeData
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                  description: Content to be analyzed
      responses:
        "200":
          description: Successful analysis
          content:
            application/json:
              schema:
                type: object
                properties:
                  tags:
                    type: array
                    items:
                      type: string
                  title:
                    type: string
                  legality:
                    type: boolean
                  description:
                    type: string
                  url:
                    type: string
                  jobId:
                    type: string
        "400":
          description: Invalid request
          content:
            application/problem+json:
              schema:
                type: object
                properties:
                  error:
                    type: string

  /status:
    get:
      summary: Check analysis status
      operationId: analyzationStatus
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: query
          required: true
          schema:
            type: string
          description: Job ID to check status for
      responses:
        "200":
          description: Successful status check
          content:
            application/json:
              schema:
                type: boolean
        "400":
          description: Invalid request
          content:
            application/problem+json:
              schema:
                type: object
                properties:
                  error:
                    type: string

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key